requirements.txt
django-allauth==0.63.3
django-crispy-forms==2.2
django-cron==0.6.0
django-crontab==0.7.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
crispy-bootstrap4==2024.1
requests==2.31.0